+++ packages/pluggableWidgets/barcode-scanner-web/src/BarcodeScanner.editorConfig.ts
@@ -1,6 +1,7 @@
 import { Properties, StructurePreviewProps, transformGroupsIntoTabs } from "@mendix/piw-utils-internal";
 import { BarcodeScannerContainerProps } from "../typings/BarcodeScannerProps";
 import BarcodeScannerSvg from "./assets/barcodescanner.svg";
+import BarcodeScannerSvgDark from "./assets/barcodescanner-dark.svg";
 
 export function getProperties(
     _: BarcodeScannerContainerProps,
@@ -13,10 +14,12 @@ export function getProperties(
     return defaultProperties;
 }
 
-export function getPreview(): StructurePreviewProps | null {
+export function getPreview(_: StructurePreviewProps, isDarkMode: boolean): StructurePreviewProps | null {
     return {
         type: "Image",
-        document: decodeURIComponent(BarcodeScannerSvg.replace("data:image/svg+xml,", "")),
+        document: decodeURIComponent(
+            (isDarkMode ? BarcodeScannerSvgDark : BarcodeScannerSvg).replace("data:image/svg+xml,", "")
+        ),
         height: 275,
         width: 275
     };