+++ packages/pluggableWidgets/barcode-scanner-web/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+- We added icons for dark mode.
+- We added dark mode to Structure mode preview.
+
 ## [2.0.1] - 2021-12-3
 
 ### Fixed